+++ CHANGELOG.md
@@ -7,7 +7,7 @@
 - Allow provider variable interpolation by deferring client instanation, `providerConfigure` only returns a configuration struct.
 
 ### Added
--
+- Composable Index Template resource, available in ESv7.8+
 
 ## [1.5.0] - 2020-10-26
 ### Changed